requirements.txt
# Core ML and AI libraries
torch>=2.0.0
transformers>=4.35.0
langchain>=0.1.0
langgraph>=0.0.40
langchain-community>=0.0.20

# Knowledge Graph and NLP
spacy>=3.7.0
jieba>=0.42.1
pkuseg>=0.0.25
networkx>=3.2.0
neo4j>=5.15.0

# Database
pymysql>=1.1.0
sqlalchemy>=2.0.0

# Web Framework
fastapi>=0.104.0
uvicorn>=0.24.0
jinja2>=3.1.0
python-multipart>=0.0.6
pydantic>=2.5.0

# Data Processing
pandas>=2.1.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Training and Evaluation
datasets>=2.14.0
accelerate>=0.24.0
peft>=0.6.0
trl>=0.7.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
tqdm>=4.66.0

# API clients
openai>=1.3.0
dashscope>=1.14.0

# RAG and Embeddings
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
chromadb>=0.4.0

# Chinese NLP
stanza>=1.7.0

# Additional utilities
loguru>=0.7.0

